LayoutTests/media/airplay-autoplay-expected.txt


** set video.src
RUN(video.src = findMediaFile('video', 'content/test'))

EVENT(canplaythrough)

** simulate choosing a device from the menu
RUN(video.webkitShowPlaybackTargetPicker())
EVENT(webkitcurrentplaybacktargetiswirelesschanged)

** start playing
RUN(video.play())

EVENT(ended)

** reset video source
RUN(video.src = findMediaFile('video', 'content/test'))
RUN(video.play())

EVENT(webkitcurrentplaybacktargetiswirelesschanged)
END OF TEST
